requirements.txt
pillow==10.4.0
PuLP==2.9.0
